Rename component.
